Revert "ExtensionDescriptorValidator"
This reverts commit c6531bfaf212bae53d4ba6484ac7ce4795edf0f8.
